Refactor kv cache controller to support different setup modes
